+++ packages/google_maps_flutter/google_maps_flutter_android/lib/src/messages.g.dart
@@ -67,6 +67,12 @@ enum PlatformPatternItemType {
   gap,
 }
 
+/// Pigeon equivalent of [MapBitmapScaling].
+enum PlatformMapBitmapScaling {
+  auto,
+  none,
+}
+
 /// Pigeon representatation of a CameraPosition.
 class PlatformCameraPosition {
   PlatformCameraPosition({
@@ -266,7 +272,7 @@ class PlatformCameraUpdateZoomBy {
 
   double amount;
 
-  PlatformOffset? focus;
+  PlatformDoublePair? focus;
 
   Object encode() {
     return <Object?>[
@@ -279,7 +285,7 @@ class PlatformCameraUpdateZoomBy {
     result as List<Object?>;
     return PlatformCameraUpdateZoomBy(
       amount: result[0]! as double,
-      focus: result[1] as PlatformOffset?,
+      focus: result[1] as PlatformDoublePair?,
     );
   }
 }
@@ -437,29 +443,29 @@ class PlatformClusterManager {
   }
 }
 
-/// Pigeon equivalent of the Offset class.
-class PlatformOffset {
-  PlatformOffset({
-    required this.dx,
-    required this.dy,
+/// Pair of double values, such as for an offset or size.
+class PlatformDoublePair {
+  PlatformDoublePair({
+    required this.x,
+    required this.y,
   });
 
-  double dx;
+  double x;
 
-  double dy;
+  double y;
 
   Object encode() {
     return <Object?>[
-      dx,
-      dy,
+      x,
+      y,
     ];
   }
 
-  static PlatformOffset decode(Object result) {
+  static PlatformDoublePair decode(Object result) {
     result as List<Object?>;
-    return PlatformOffset(
-      dx: result[0]! as double,
-      dy: result[1]! as double,
+    return PlatformDoublePair(
+      x: result[0]! as double,
+      y: result[1]! as double,
     );
   }
 }
@@ -476,7 +482,7 @@ class PlatformInfoWindow {
 
   String? snippet;
 
-  PlatformOffset anchor;
+  PlatformDoublePair anchor;
 
   Object encode() {
     return <Object?>[
@@ -491,7 +497,7 @@ class PlatformInfoWindow {
     return PlatformInfoWindow(
       title: result[0] as String?,
       snippet: result[1] as String?,
-      anchor: result[2]! as PlatformOffset,
+      anchor: result[2]! as PlatformDoublePair,
     );
   }
 }
@@ -504,7 +510,7 @@ class PlatformMarker {
     this.consumeTapEvents = false,
     this.draggable = false,
     this.flat = false,
-    this.icon = const <Object>['defaultMarker'],
+    required this.icon,
     required this.infoWindow,
     required this.position,
     this.rotation = 0.0,
@@ -516,7 +522,7 @@ class PlatformMarker {
 
   double alpha;
 
-  PlatformOffset anchor;
+  PlatformDoublePair anchor;
 
   bool consumeTapEvents;
 
@@ -524,8 +530,7 @@ class PlatformMarker {
 
   bool flat;
 
-  /// The icon as JSON data.
-  Object icon;
+  PlatformBitmap icon;
 
   PlatformInfoWindow infoWindow;
 
@@ -563,11 +568,11 @@ class PlatformMarker {
     result as List<Object?>;
     return PlatformMarker(
       alpha: result[0]! as double,
-      anchor: result[1]! as PlatformOffset,
+      anchor: result[1]! as PlatformDoublePair,
       consumeTapEvents: result[2]! as bool,
       draggable: result[3]! as bool,
       flat: result[4]! as bool,
-      icon: result[5]!,
+      icon: result[5]! as PlatformBitmap,
       infoWindow: result[6]! as PlatformInfoWindow,
       position: result[7]! as PlatformLatLng,
       rotation: result[8]! as double,
@@ -738,8 +743,7 @@ class PlatformCap {
 
   PlatformCapType type;
 
-  /// The JSON data returned by BitmapDescriptor.toJson.
-  Object? bitmapDescriptor;
+  PlatformBitmap? bitmapDescriptor;
 
   double? refWidth;
 
@@ -755,7 +759,7 @@ class PlatformCap {
     result as List<Object?>;
     return PlatformCap(
       type: result[0]! as PlatformCapType,
-      bitmapDescriptor: result[1],
+      bitmapDescriptor: result[1] as PlatformBitmap?,
       refWidth: result[2] as double?,
     );
   }
@@ -1293,6 +1297,235 @@ class PlatformZoomRange {
   }
 }
 
+/// Pigeon equivalent of [BitmapDescriptor]. As there are multiple disjoint
+/// types of [BitmapDescriptor], [PlatformBitmap] contains a single field which
+/// may hold the pigeon equivalent type of any of them.
+class PlatformBitmap {
+  PlatformBitmap({
+    required this.bitmap,
+  });
+
+  /// One of [PlatformBitmapAssetMap], [PlatformBitmapAsset],
+  /// [PlatformBitmapAssetImage], [PlatformBitmapBytesMap],
+  /// [PlatformBitmapBytes], or [PlatformBitmapDefaultMarker].
+  /// As Pigeon does not currently support data class inheritance, this
+  /// approach allows for the different bitmap implementations to be valid
+  /// argument and return types of the API methods. See
+  /// https://github.com/flutter/flutter/issues/117819.
+  Object bitmap;
+
+  Object encode() {
+    return <Object?>[
+      bitmap,
+    ];
+  }
+
+  static PlatformBitmap decode(Object result) {
+    result as List<Object?>;
+    return PlatformBitmap(
+      bitmap: result[0]!,
+    );
+  }
+}
+
+/// Pigeon equivalent of [DefaultMarker]. See
+/// https://developers.google.com/maps/documentation/android-sdk/reference/com/google/android/libraries/maps/model/BitmapDescriptorFactory#defaultMarker(float)
+class PlatformBitmapDefaultMarker {
+  PlatformBitmapDefaultMarker({
+    this.hue,
+  });
+
+  double? hue;
+
+  Object encode() {
+    return <Object?>[
+      hue,
+    ];
+  }
+
+  static PlatformBitmapDefaultMarker decode(Object result) {
+    result as List<Object?>;
+    return PlatformBitmapDefaultMarker(
+      hue: result[0] as double?,
+    );
+  }
+}
+
+/// Pigeon equivalent of [BytesBitmap]. See
+/// https://developers.google.com/maps/documentation/android-sdk/reference/com/google/android/libraries/maps/model/BitmapDescriptorFactory#fromBitmap(android.graphics.Bitmap)
+class PlatformBitmapBytes {
+  PlatformBitmapBytes({
+    required this.byteData,
+    this.size,
+  });
+
+  Uint8List byteData;
+
+  PlatformDoublePair? size;
+
+  Object encode() {
+    return <Object?>[
+      byteData,
+      size,
+    ];
+  }
+
+  static PlatformBitmapBytes decode(Object result) {
+    result as List<Object?>;
+    return PlatformBitmapBytes(
+      byteData: result[0]! as Uint8List,
+      size: result[1] as PlatformDoublePair?,
+    );
+  }
+}
+
+/// Pigeon equivalent of [AssetBitmap]. See
+/// https://developers.google.com/maps/documentation/android-sdk/reference/com/google/android/libraries/maps/model/BitmapDescriptorFactory#public-static-bitmapdescriptor-fromasset-string-assetname
+class PlatformBitmapAsset {
+  PlatformBitmapAsset({
+    required this.name,
+    this.pkg,
+  });
+
+  String name;
+
+  String? pkg;
+
+  Object encode() {
+    return <Object?>[
+      name,
+      pkg,
+    ];
+  }
+
+  static PlatformBitmapAsset decode(Object result) {
+    result as List<Object?>;
+    return PlatformBitmapAsset(
+      name: result[0]! as String,
+      pkg: result[1] as String?,
+    );
+  }
+}
+
+/// Pigeon equivalent of [AssetImageBitmap]. See
+/// https://developers.google.com/maps/documentation/android-sdk/reference/com/google/android/libraries/maps/model/BitmapDescriptorFactory#public-static-bitmapdescriptor-fromasset-string-assetname
+class PlatformBitmapAssetImage {
+  PlatformBitmapAssetImage({
+    required this.name,
+    required this.scale,
+    this.size,
+  });
+
+  String name;
+
+  double scale;
+
+  PlatformDoublePair? size;
+
+  Object encode() {
+    return <Object?>[
+      name,
+      scale,
+      size,
+    ];
+  }
+
+  static PlatformBitmapAssetImage decode(Object result) {
+    result as List<Object?>;
+    return PlatformBitmapAssetImage(
+      name: result[0]! as String,
+      scale: result[1]! as double,
+      size: result[2] as PlatformDoublePair?,
+    );
+  }
+}
+
+/// Pigeon equivalent of [AssetMapBitmap]. See
+/// https://developers.google.com/maps/documentation/android-sdk/reference/com/google/android/libraries/maps/model/BitmapDescriptorFactory#public-static-bitmapdescriptor-fromasset-string-assetname
+class PlatformBitmapAssetMap {
+  PlatformBitmapAssetMap({
+    required this.assetName,
+    required this.bitmapScaling,
+    required this.imagePixelRatio,
+    this.width,
+    this.height,
+  });
+
+  String assetName;
+
+  PlatformMapBitmapScaling bitmapScaling;
+
+  double imagePixelRatio;
+
+  double? width;
+
+  double? height;
+
+  Object encode() {
+    return <Object?>[
+      assetName,
+      bitmapScaling,
+      imagePixelRatio,
+      width,
+      height,
+    ];
+  }
+
+  static PlatformBitmapAssetMap decode(Object result) {
+    result as List<Object?>;
+    return PlatformBitmapAssetMap(
+      assetName: result[0]! as String,
+      bitmapScaling: result[1]! as PlatformMapBitmapScaling,
+      imagePixelRatio: result[2]! as double,
+      width: result[3] as double?,
+      height: result[4] as double?,
+    );
+  }
+}
+
+/// Pigeon equivalent of [BytesMapBitmap]. See
+/// https://developers.google.com/maps/documentation/android-sdk/reference/com/google/android/libraries/maps/model/BitmapDescriptorFactory#public-static-bitmapdescriptor-frombitmap-bitmap-image
+class PlatformBitmapBytesMap {
+  PlatformBitmapBytesMap({
+    required this.byteData,
+    required this.bitmapScaling,
+    required this.imagePixelRatio,
+    this.width,
+    this.height,
+  });
+
+  Uint8List byteData;
+
+  PlatformMapBitmapScaling bitmapScaling;
+
+  double imagePixelRatio;
+
+  double? width;
+
+  double? height;
+
+  Object encode() {
+    return <Object?>[
+      byteData,
+      bitmapScaling,
+      imagePixelRatio,
+      width,
+      height,
+    ];
+  }
+
+  static PlatformBitmapBytesMap decode(Object result) {
+    result as List<Object?>;
+    return PlatformBitmapBytesMap(
+      byteData: result[0]! as Uint8List,
+      bitmapScaling: result[1]! as PlatformMapBitmapScaling,
+      imagePixelRatio: result[2]! as double,
+      width: result[3] as double?,
+      height: result[4] as double?,
+    );
+  }
+}
+
 class _PigeonCodec extends StandardMessageCodec {
   const _PigeonCodec();
   @override
@@ -1315,101 +1548,125 @@ class _PigeonCodec extends StandardMessageCodec {
     } else if (value is PlatformPatternItemType) {
       buffer.putUint8(133);
       writeValue(buffer, value.index);
-    } else if (value is PlatformCameraPosition) {
+    } else if (value is PlatformMapBitmapScaling) {
       buffer.putUint8(134);
-      writeValue(buffer, value.encode());
-    } else if (value is PlatformCameraUpdate) {
+      writeValue(buffer, value.index);
+    } else if (value is PlatformCameraPosition) {
       buffer.putUint8(135);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformCameraUpdateNewCameraPosition) {
+    } else if (value is PlatformCameraUpdate) {
       buffer.putUint8(136);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformCameraUpdateNewLatLng) {
+    } else if (value is PlatformCameraUpdateNewCameraPosition) {
       buffer.putUint8(137);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformCameraUpdateNewLatLngBounds) {
+    } else if (value is PlatformCameraUpdateNewLatLng) {
       buffer.putUint8(138);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformCameraUpdateNewLatLngZoom) {
+    } else if (value is PlatformCameraUpdateNewLatLngBounds) {
       buffer.putUint8(139);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformCameraUpdateScrollBy) {
+    } else if (value is PlatformCameraUpdateNewLatLngZoom) {
       buffer.putUint8(140);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformCameraUpdateZoomBy) {
+    } else if (value is PlatformCameraUpdateScrollBy) {
       buffer.putUint8(141);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformCameraUpdateZoom) {
+    } else if (value is PlatformCameraUpdateZoomBy) {
       buffer.putUint8(142);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformCameraUpdateZoomTo) {
+    } else if (value is PlatformCameraUpdateZoom) {
       buffer.putUint8(143);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformCircle) {
+    } else if (value is PlatformCameraUpdateZoomTo) {
       buffer.putUint8(144);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformHeatmap) {
+    } else if (value is PlatformCircle) {
       buffer.putUint8(145);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformClusterManager) {
+    } else if (value is PlatformHeatmap) {
       buffer.putUint8(146);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformOffset) {
+    } else if (value is PlatformClusterManager) {
       buffer.putUint8(147);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformInfoWindow) {
+    } else if (value is PlatformDoublePair) {
       buffer.putUint8(148);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformMarker) {
+    } else if (value is PlatformInfoWindow) {
       buffer.putUint8(149);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformPolygon) {
+    } else if (value is PlatformMarker) {
       buffer.putUint8(150);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformPolyline) {
+    } else if (value is PlatformPolygon) {
       buffer.putUint8(151);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformCap) {
+    } else if (value is PlatformPolyline) {
       buffer.putUint8(152);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformPatternItem) {
+    } else if (value is PlatformCap) {
       buffer.putUint8(153);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformTile) {
+    } else if (value is PlatformPatternItem) {
       buffer.putUint8(154);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformTileOverlay) {
+    } else if (value is PlatformTile) {
       buffer.putUint8(155);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformEdgeInsets) {
+    } else if (value is PlatformTileOverlay) {
       buffer.putUint8(156);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformLatLng) {
+    } else if (value is PlatformEdgeInsets) {
       buffer.putUint8(157);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformLatLngBounds) {
+    } else if (value is PlatformLatLng) {
       buffer.putUint8(158);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformCluster) {
+    } else if (value is PlatformLatLngBounds) {
       buffer.putUint8(159);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformCameraTargetBounds) {
+    } else if (value is PlatformCluster) {
       buffer.putUint8(160);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformMapViewCreationParams) {
+    } else if (value is PlatformCameraTargetBounds) {
       buffer.putUint8(161);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformMapConfiguration) {
+    } else if (value is PlatformMapViewCreationParams) {
       buffer.putUint8(162);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformPoint) {
+    } else if (value is PlatformMapConfiguration) {
       buffer.putUint8(163);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformTileLayer) {
+    } else if (value is PlatformPoint) {
       buffer.putUint8(164);
       writeValue(buffer, value.encode());
-    } else if (value is PlatformZoomRange) {
+    } else if (value is PlatformTileLayer) {
       buffer.putUint8(165);
+      writeValue(buffer, value.encode());
+    } else if (value is PlatformZoomRange) {
+      buffer.putUint8(166);
+      writeValue(buffer, value.encode());
+    } else if (value is PlatformBitmap) {
+      buffer.putUint8(167);
+      writeValue(buffer, value.encode());
+    } else if (value is PlatformBitmapDefaultMarker) {
+      buffer.putUint8(168);
+      writeValue(buffer, value.encode());
+    } else if (value is PlatformBitmapBytes) {
+      buffer.putUint8(169);
+      writeValue(buffer, value.encode());
+    } else if (value is PlatformBitmapAsset) {
+      buffer.putUint8(170);
+      writeValue(buffer, value.encode());
+    } else if (value is PlatformBitmapAssetImage) {
+      buffer.putUint8(171);
+      writeValue(buffer, value.encode());
+    } else if (value is PlatformBitmapAssetMap) {
+      buffer.putUint8(172);
+      writeValue(buffer, value.encode());
+    } else if (value is PlatformBitmapBytesMap) {
+      buffer.putUint8(173);
       writeValue(buffer, value.encode());
     } else {
       super.writeValue(buffer, value);
@@ -1435,69 +1692,86 @@ class _PigeonCodec extends StandardMessageCodec {
         final int? value = readValue(buffer) as int?;
         return value == null ? null : PlatformPatternItemType.values[value];
       case 134:
-        return PlatformCameraPosition.decode(readValue(buffer)!);
+        final int? value = readValue(buffer) as int?;
+        return value == null ? null : PlatformMapBitmapScaling.values[value];
       case 135:
-        return PlatformCameraUpdate.decode(readValue(buffer)!);
+        return PlatformCameraPosition.decode(readValue(buffer)!);
       case 136:
-        return PlatformCameraUpdateNewCameraPosition.decode(readValue(buffer)!);
+        return PlatformCameraUpdate.decode(readValue(buffer)!);
       case 137:
-        return PlatformCameraUpdateNewLatLng.decode(readValue(buffer)!);
+        return PlatformCameraUpdateNewCameraPosition.decode(readValue(buffer)!);
       case 138:
-        return PlatformCameraUpdateNewLatLngBounds.decode(readValue(buffer)!);
+        return PlatformCameraUpdateNewLatLng.decode(readValue(buffer)!);
       case 139:
-        return PlatformCameraUpdateNewLatLngZoom.decode(readValue(buffer)!);
+        return PlatformCameraUpdateNewLatLngBounds.decode(readValue(buffer)!);
       case 140:
-        return PlatformCameraUpdateScrollBy.decode(readValue(buffer)!);
+        return PlatformCameraUpdateNewLatLngZoom.decode(readValue(buffer)!);
       case 141:
-        return PlatformCameraUpdateZoomBy.decode(readValue(buffer)!);
+        return PlatformCameraUpdateScrollBy.decode(readValue(buffer)!);
       case 142:
-        return PlatformCameraUpdateZoom.decode(readValue(buffer)!);
+        return PlatformCameraUpdateZoomBy.decode(readValue(buffer)!);
       case 143:
-        return PlatformCameraUpdateZoomTo.decode(readValue(buffer)!);
+        return PlatformCameraUpdateZoom.decode(readValue(buffer)!);
       case 144:
-        return PlatformCircle.decode(readValue(buffer)!);
+        return PlatformCameraUpdateZoomTo.decode(readValue(buffer)!);
       case 145:
-        return PlatformHeatmap.decode(readValue(buffer)!);
+        return PlatformCircle.decode(readValue(buffer)!);
       case 146:
-        return PlatformClusterManager.decode(readValue(buffer)!);
+        return PlatformHeatmap.decode(readValue(buffer)!);
       case 147:
-        return PlatformOffset.decode(readValue(buffer)!);
+        return PlatformClusterManager.decode(readValue(buffer)!);
       case 148:
-        return PlatformInfoWindow.decode(readValue(buffer)!);
+        return PlatformDoublePair.decode(readValue(buffer)!);
       case 149:
-        return PlatformMarker.decode(readValue(buffer)!);
+        return PlatformInfoWindow.decode(readValue(buffer)!);
       case 150:
-        return PlatformPolygon.decode(readValue(buffer)!);
+        return PlatformMarker.decode(readValue(buffer)!);
       case 151:
-        return PlatformPolyline.decode(readValue(buffer)!);
+        return PlatformPolygon.decode(readValue(buffer)!);
       case 152:
-        return PlatformCap.decode(readValue(buffer)!);
+        return PlatformPolyline.decode(readValue(buffer)!);
       case 153:
-        return PlatformPatternItem.decode(readValue(buffer)!);
+        return PlatformCap.decode(readValue(buffer)!);
       case 154:
-        return PlatformTile.decode(readValue(buffer)!);
+        return PlatformPatternItem.decode(readValue(buffer)!);
       case 155:
-        return PlatformTileOverlay.decode(readValue(buffer)!);
+        return PlatformTile.decode(readValue(buffer)!);
       case 156:
-        return PlatformEdgeInsets.decode(readValue(buffer)!);
+        return PlatformTileOverlay.decode(readValue(buffer)!);
       case 157:
-        return PlatformLatLng.decode(readValue(buffer)!);
+        return PlatformEdgeInsets.decode(readValue(buffer)!);
       case 158:
-        return PlatformLatLngBounds.decode(readValue(buffer)!);
+        return PlatformLatLng.decode(readValue(buffer)!);
       case 159:
-        return PlatformCluster.decode(readValue(buffer)!);
+        return PlatformLatLngBounds.decode(readValue(buffer)!);
       case 160:
-        return PlatformCameraTargetBounds.decode(readValue(buffer)!);
+        return PlatformCluster.decode(readValue(buffer)!);
       case 161:
-        return PlatformMapViewCreationParams.decode(readValue(buffer)!);
+        return PlatformCameraTargetBounds.decode(readValue(buffer)!);
       case 162:
-        return PlatformMapConfiguration.decode(readValue(buffer)!);
+        return PlatformMapViewCreationParams.decode(readValue(buffer)!);
       case 163:
-        return PlatformPoint.decode(readValue(buffer)!);
+        return PlatformMapConfiguration.decode(readValue(buffer)!);
       case 164:
-        return PlatformTileLayer.decode(readValue(buffer)!);
+        return PlatformPoint.decode(readValue(buffer)!);
       case 165:
+        return PlatformTileLayer.decode(readValue(buffer)!);
+      case 166:
         return PlatformZoomRange.decode(readValue(buffer)!);
+      case 167:
+        return PlatformBitmap.decode(readValue(buffer)!);
+      case 168:
+        return PlatformBitmapDefaultMarker.decode(readValue(buffer)!);
+      case 169:
+        return PlatformBitmapBytes.decode(readValue(buffer)!);
+      case 170:
+        return PlatformBitmapAsset.decode(readValue(buffer)!);
+      case 171:
+        return PlatformBitmapAssetImage.decode(readValue(buffer)!);
+      case 172:
+        return PlatformBitmapAssetMap.decode(readValue(buffer)!);
+      case 173:
+        return PlatformBitmapBytesMap.decode(readValue(buffer)!);
       default:
         return super.readValueOfType(type, buffer);
     }