+++ packages/interactive_media_ads/android/src/main/kotlin/dev/flutter/packages/interactive_media_ads/ProxyApiRegistrar.kt
@@ -137,4 +137,12 @@ open class ProxyApiRegistrar(binaryMessenger: BinaryMessenger, var context: Cont
   override fun getPigeonApiCompanionAd(): PigeonApiCompanionAd {
     return CompanionAdProxyApi(this)
   }
+
+  override fun getPigeonApiUniversalAdId(): PigeonApiUniversalAdId {
+    return UniversalAdIdProxyApi(this)
+  }
+
+  override fun getPigeonApiAd(): PigeonApiAd {
+    return AdProxyApi(this)
+  }
 }