+++ packages/interactive_media_ads/pigeons/interactive_media_ads_ios.dart
@@ -2,12 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// TODO(bparrishMines): Uncomment this file once
-// https://github.com/flutter/packages/pull/6602 lands. This file uses the
-// Swift ProxyApi feature from pigeon.
 // ignore_for_file: avoid_unused_constructor_parameters
 
-/*
 import 'package:pigeon/pigeon.dart';
 
 @ConfigurePigeon(
@@ -573,7 +569,7 @@ abstract class NSObject {}
 ///
 /// See https://developers.google.com/ad-manager/dynamic-ad-insertion/sdk/ios/reference/Classes/IMAFriendlyObstruction.html.
 @ProxyApi()
-abstract class IMAFriendlyObstruction {
+abstract class IMAFriendlyObstruction extends NSObject {
   /// Initializes a friendly obstruction.
   IMAFriendlyObstruction();
 
@@ -590,4 +586,25 @@ abstract class IMAFriendlyObstruction {
   /// spaces.
   late final String? detailedReason;
 }
-*/
+
+/// An object that holds data corresponding to the companion ad.
+///
+/// See https://developers.google.com/interactive-media-ads/docs/sdks/ios/client-side/reference/Classes/IMACompanionAd.
+@ProxyApi()
+abstract class IMACompanionAd extends NSObject {
+  /// The value for the resource of this companion.
+  late final String? resourceValue;
+
+  /// The API needed to execute this ad, or nil if unavailable.
+  late final String? apiFramework;
+
+  /// The width of the companion in pixels.
+  ///
+  /// 0 if unavailable.
+  late final int width;
+
+  /// The height of the companion in pixels.
+  ///
+  /// 0 if unavailable.
+  late final int height;
+}